# services/orchestrator-api/openapi.yaml
openapi: 3.0.3
info:
  title: Orchestrator API
  version: "1.0.0"
  description: >
    API para ingestão de eventos OCPP, comandos, consulta de sessões, métricas,
    tarifas e faturamento.
servers:
  - url: http://localhost:3000
    description: Local
...
security:
  - ApiKeyAuth: []   # por padrão todas as rotas exigem X-API-Key
paths:
...
  /v1/billing/invoices:
    get:
      tags: [Billing]
      summary: Lista de faturas
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [issued, paid, cancelled]
        - in: query
          name: charge_box_id
          schema: { type: string, nullable: true }
        - in: query
          name: from
          schema: { type: string, format: date-time, nullable: true }
        - in: query
          name: to
          schema: { type: string, format: date-time, nullable: true }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Lista de faturas
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/InvoiceSummary' }
  /v1/billing/invoices/{id}:
    get:
      tags: [Billing]
      summary: Detalhe da fatura
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Fatura
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invoice' }
        '404': { description: Não encontrado }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    InvoiceSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        status:
          type: string
          enum: [issued, paid, cancelled]
        amount: { type: number, format: double }
        currency: { type: string, example: BRL }
        issued_at: { type: string, format: date-time }
    Invoice:
      allOf:
        - $ref: '#/components/schemas/InvoiceSummary'
        - type: object
          properties:
            session_transaction_id: { type: integer, format: int32, nullable: true }
            charge_box_id: { type: string, nullable: true }
            id_tag: { type: string, nullable: true }
            energy_kwh: { type: number, format: double, nullable: true }
            due_at: { type: string, format: date-time, nullable: true }
            paid_at: { type: string, format: date-time, nullable: true }
            metadata:
              type: object
              additionalProperties: true